+++ src/Exceptions/ShouldBeSkippedMetricExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace AvtoDev\AppMetrics\Exceptions;
+
+use Throwable;
+
+interface ShouldBeSkippedMetricExceptionInterface extends Throwable
+{
+    //
+}